// ANGULAR/ANGULAR/src/app/Shared/Enum/scheme.enum.ts
export enum SchemeTypeEnum {
  Both = '28',
  Group = 1,
  Individual = 2,
  Scheme = '20062',
  Logo_width = '180',
  Banner_height = '490',
  Banner_width = '1920'

}

export enum MPRMaxCount {
  Count = 15,
}

export enum GalleryUploadType {
  Photo = "PHOTO",
  Video = "VIDEO",
  URL = "URL",
  PDF = "PDF",
  VIDEO_URL = "VIDEO URL"
}

export enum ContactPersionTypeEnum {
  StateLabel = 1,
  DistrictLabel = 2,
  BlockLabel = 3
}

export enum FileType {
  Photo = "image.*",
  Video = "video.*",
  PDF = "application/pdf"
}

export enum StatusEnum {
  Active = 1,
  DeActive = 0,
  Both = -1
}